fix: Rearrange the reporter package a little bit
