+++ reporting/constants.go
@@ -1,0 +1,5 @@
+package reporting
+
+const DEFAULT_SLACK_ICON = " "
+const NO_OWNER_KEY = "__none__"
+const SUMMARY_KEY = "summary"
